chore: cleanup openapi schema tests
